dailyViewChange 분리 및 daily total 안바뀌는 버그 해결
